ags: Proof-of-concept of a translation selection from command line argument
If a language is requested, select a matching list of words which appear
in typical translation files for that language. Then, find if any matches
any translation file at all. If it does, pick that file as a translation.
If no file is found or no language was requested, look for a pre-configured
translation, as before this change.